+++ pdi/src/global_context.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2015-2025 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2015-2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
  * Copyright (C) 2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
@@ -34,7 +34,6 @@
 #include <unordered_map>
 
 #include "pdi/pdi_fwd.h"
-#include "pdi/callbacks.h"
 #include "pdi/context.h"
 #include "pdi/context_proxy.h"
 #include "pdi/data_descriptor.h"
@@ -45,7 +44,7 @@
 #include "plugin_store.h"
 
 namespace PDI {
-
+//
 class PDI_EXPORT Global_context: public Context
 {
 private:
@@ -66,12 +65,97 @@ private:
 	/// The plugins, this should be late in the list to be destroyed early
 	Plugin_store m_plugins;
 
-	/// Callbacks of the context
-	Callbacks m_callbacks;
+	/**
+	 *  Callbacks called after init
+	 *
+	 *  This must be a list, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::list<std::function<void()>> m_init_callbacks;
+
+	/**
+	 *  Callbacks called when any data is available
+	 *
+	 *  This must be a list, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::list<std::function<void(const std::string&, Ref)>> m_data_callbacks;
+
+	/**
+	 *  Callbacks called when specified data is available.
+	 *
+	 *  This must be an ordered multimap, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::multimap<std::string, std::function<void(const std::string&, Ref)>> m_named_data_callbacks;
+
+	/**
+	 *  Callbacks called when any data is reclaimed/released.
+	 *
+	 *  This must be a list, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::list<std::function<void(const std::string&, Ref)>> m_data_remove_callbacks;
+
+	/**
+	 *  Callbacks called when specified data is reclaimed/released.
+	 *
+	 *  This must be an ordered multimap, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::multimap<std::string, std::function<void(const std::string&, Ref)>> m_named_data_remove_callbacks;
+
+	/**
+	 *  Callbacks called on any event
+	 *
+	 *  This must be a list, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::list<std::function<void(const std::string&)>> m_event_callbacks;
+
+	/**
+	 *  Callbacks called on specified event
+	 *
+	 *  This must be an ordered multimap, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::multimap<std::string, std::function<void(const std::string&)>> m_named_event_callbacks;
+
+	/**
+	 *  Callbacks called on any empty desc access
+	 *
+	 *  This must be a list, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::list<std::function<void(const std::string&)>> m_empty_desc_access_callbacks;
+
+	/**
+	 *  Callbacks called on specified empty desc access
+	 *
+	 *  This must be an ordered multimap, because valid iterators are needed to properly remove the callback by plugin
+	 */
+	std::multimap<std::string, std::function<void(const std::string&)>> m_named_empty_desc_access_callbacks;
 
 	Global_context(const Global_context&) = delete;
 
 	Global_context(Global_context&&) = delete;
+
+	/// Calls init callbacks
+	void notify_init() const;
+
+	/** Calls data callbacks
+	 *  \param name name of the shared descriptor
+	 *  \param ref shared reference
+	 */
+	void notify_data(const std::string& name, Ref ref);
+
+	/** Calls data remove callbacks
+	 *  \param name name of the descriptor that will be reclaimed/released
+	 *  \param ref reference that will be reclaimed/released
+	 */
+	void notify_data_remove(const std::string& name, Ref ref);
+
+	/** Calls event callbacks
+	 *  \param name name of the event
+	 */
+	void notify_event(const std::string& name);
+
+	/** Calls missing data callbacks
+	 *  \param name name of the accessed descriptor
+	 */
+	void notify_missing_data(const std::string& name);
 
 public:
 	static void init(PC_tree_t conf);
@@ -84,37 +168,22 @@ public:
 
 	Global_context(PC_tree_t conf);
 
-	/** Accesses the descriptor for a specific name. Might be uninitialized
-	 */
+	~Global_context() override;
+
 	Data_descriptor& desc(const std::string& name) override;
 
-	/** Accesses the descriptor for a specific name. Might be uninitialized
-	 */
 	Data_descriptor& desc(const char* name) override;
 
-	/** Accesses the descriptor for a specific name. Might be uninitialized
-	 */
 	Data_descriptor& operator[] (const std::string& name) override;
 
-	/** Accesses the descriptor for a specific name. Might be uninitialized
-	 */
 	Data_descriptor& operator[] (const char* name) override;
 
-	/** Returns an iterator on the first descriptor
-	 */
 	Iterator begin() override;
 
-	/** Returns an iterator past the last descriptor
-	 */
 	Iterator end() override;
 
-	/** Returns the iterator that corresponds to the provided Data_descriptor name
-	 */
 	Iterator find(const std::string& name) override;
 
-	/** Triggers a PDI "event"
-	 * \param[in] name the event name
-	 */
 	void event(const char* name) override;
 
 	Logger& logger() override;
@@ -123,11 +192,15 @@ public:
 
 	void add_datatype(const std::string&, Datatype_template_parser) override;
 
-	Callbacks& callbacks() override;
+	std::function<void()> on_init(const std::function<void()>& callback) override;
 
-	void finalize_and_exit() override;
+	std::function<void()> on_data(const std::function<void(const std::string&, Ref)>& callback, const std::string& name = {}) override;
 
-	~Global_context() override;
+	std::function<void()> on_data_remove(const std::function<void(const std::string&, Ref)>& callback, const std::string& name = {}) override;
+
+	std::function<void()> on_event(const std::function<void(const std::string&)>& callback, const std::string& name = {}) override;
+
+	std::function<void()> on_missing_data(const std::function<void(const std::string&)>& callback, const std::string& name = {}) override;
 };
 
 } // namespace PDI