+++ plugins/serialize/serialize.cxx
@@ -291,10 +291,8 @@ struct serialize_plugin: PDI::Plugin {
 
 			context().logger().trace("Sharing `{}' PDI_INOUT", serialized_name);
 			context().desc(serialized_name).share(serialized_ref, false, false);
-			std::function<void()> remove_callback = context().callbacks().add_data_remove_callback(
-				[this](const std::string& desc_name, PDI::Ref ref) { release_serialized(desc_name, ref); },
-				desc_name
-			);
+			std::function<void()> remove_callback
+				= context().on_data_remove([this](const std::string& desc_name, PDI::Ref ref) { release_serialized(desc_name, ref); }, desc_name);
 			m_serialized_remove_callback.emplace_back(serialized_name, remove_callback, PDI_INOUT);
 
 		} else if (PDI::Ref_r ref_r = ref) {
@@ -312,10 +310,8 @@ struct serialize_plugin: PDI::Plugin {
 			}
 			context().logger().trace("Sharing `{}' PDI_OUT", serialized_name);
 			context().desc(serialized_name).share(serialized_ref, true, false);
-			std::function<void()> remove_callback = context().callbacks().add_data_remove_callback(
-				[this](const std::string& desc_name, PDI::Ref ref) { release_serialized(desc_name, ref); },
-				desc_name
-			);
+			std::function<void()> remove_callback
+				= context().on_data_remove([this](const std::string& desc_name, PDI::Ref ref) { release_serialized(desc_name, ref); }, desc_name);
 			m_serialized_remove_callback.emplace_back(serialized_name, remove_callback, PDI_OUT);
 
 		} else if (PDI::Ref_w ref_w{ref}) {
@@ -327,10 +323,8 @@ struct serialize_plugin: PDI::Plugin {
 
 			context().logger().trace("Sharing `{}' PDI_IN", serialized_name);
 			context().desc(serialized_name).share(serialized_ref, false, false);
-			std::function<void()> remove_callback = context().callbacks().add_data_remove_callback(
-				[this](const std::string& desc_name, PDI::Ref ref) { release_serialized(desc_name, ref); },
-				desc_name
-			);
+			std::function<void()> remove_callback
+				= context().on_data_remove([this](const std::string& desc_name, PDI::Ref ref) { release_serialized(desc_name, ref); }, desc_name);
 			m_serialized_remove_callback.emplace_back(serialized_name, remove_callback, PDI_IN);
 		}
 	}
@@ -388,10 +382,7 @@ struct serialize_plugin: PDI::Plugin {
 			std::string desc_name = PDI::to_string(key);
 			m_desc_to_serialize.emplace(desc_name, PDI::to_string(value));
 			context().logger().trace("`{}' will be serialized", desc_name);
-			context().callbacks().add_data_callback(
-				[this](const std::string& desc_name, PDI::Ref ref) { share_serialized(desc_name, ref); },
-				desc_name
-			);
+			context().on_data([this](const std::string& desc_name, PDI::Ref ref) { share_serialized(desc_name, ref); }, desc_name);
 		});
 	}